pyproject.toml
[project]
name = "calmtechrss"
version = "0.1.0"
description = "Generate a calm Chinese technology daily digest from RSS sources."
requires-python = ">=3.11"
dependencies = [
  "feedparser>=6.0.11",
  "httpx>=0.27.0",
  "Jinja2>=3.1.3",
  "numpy>=1.26.0",
  "python-dateutil>=2.9.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.7.0"]
dev = ["pytest>=8.0.0"]

[project.scripts]
calmtechrss = "calmtechrss.cli:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["calmtechrss*"]

[tool.setuptools.package-data]
calmtechrss = ["templates/*.j2"]

